Upgraded packages and fixed most remaining TypeScript errors
