+++ package.json
@@ -1,33 +1,50 @@
 {
-  "name": "supercharged-links-obsidian",
-  "version": "0.7.1",
+  "name": "mycoshiro-supercharged-links",
+  "version": "0.7.3",
   "description": "Add properties and menu options to links and style them!",
   "main": "main.js",
   "scripts": {
-    "dev": "rollup --config rollup.config.js -w",
-    "build": "rollup --config rollup.config.js --environment BUILD:production",
-    "release": "standard-version"
-  },
-  "standard-version": {
-    "t": ""
+    "dev": "node esbuild.config.mjs",
+    "build": "tsc -noEmit -skipLibCheck && node esbuild.config.mjs production",
+    "deploy": "cross-env-shell \"yarn shx --verbose cp -u main.js manifest.json versions.json \"${PROJECT_CWD}/.obsidian/plugins/${npm_package_name}\" \"",
+    "lint": "eslint */**.ts",
+    "lint:fix": "eslint --fix */**.ts"
   },
   "keywords": [],
-  "author": "",
+  "author": "Joel Van Eenwyk",
   "license": "MIT",
   "devDependencies": {
-    "@codemirror/language": "^6.9.0",
-    "@codemirror/state": "^6.0.0",
-    "@codemirror/view": "^6.0.0",
-    "@rollup/plugin-commonjs": "^18.0.0",
-    "@rollup/plugin-node-resolve": "^11.2.1",
-    "@rollup/plugin-typescript": "^8.2.1",
-    "@types/node": "^14.14.37",
-    "obsidian": "^0.15.0",
-    "obsidian-dataview": "^0.4.21",
-    "rollup": "^2.32.1",
-    "standard-version": "^9.3.2",
-    "tslib": "^2.2.0",
-    "typescript": "^4.2.4"
+    "@codemirror/buildhelper": "^1.0.0",
+    "@codemirror/state": "^6.2.1",
+    "@codemirror/view": "^6.16.0",
+    "@lezer/common": "^1.0.0",
+    "@lezer/highlight": "^1.0.0",
+    "@lezer/javascript": "^1.0.0",
+    "@lezer/lr": "^1.0.0",
+    "@rollup/plugin-commonjs": "^25.0.4",
+    "@rollup/plugin-node-resolve": "^15.2.0",
+    "@rollup/plugin-typescript": "^11.1.2",
+    "@types/jest": "^29.5.3",
+    "@types/mocha": "^10.0.1",
+    "@types/node": "^20.5.1",
+    "@typescript-eslint/eslint-plugin": "6.4.0",
+    "@typescript-eslint/parser": "6.4.0",
+    "builtin-modules": "3.3.0",
+    "cross-env": "^7.0.3",
+    "esbuild": "^0.19.2",
+    "eslint": "^8.47.0",
+    "ist": "^1.1.7",
+    "obsidian": "^1.4.0",
+    "obsidian-dataview": "^0.5.56",
+    "rollup": "^3.28.0",
+    "shx": "^0.3.4",
+    "standard-version": "^9.5.0",
+    "style-mod": "^4.0.3",
+    "tslib": "^2.6.2",
+    "typescript": "^5.1.6"
   },
-  "packageManager": "yarn@3.6.2"
+  "dependencies": {
+    "@codemirror/language": "^6.9.0",
+    "@yarnpkg/sdks": "^3.0.0-rc.49"
+  }
 }
