--- rollup.config.js
@@ -1,48 +1,0 @@
-import typescript from '@rollup/plugin-typescript';
-import {nodeResolve} from '@rollup/plugin-node-resolve';
-import commonjs from '@rollup/plugin-commonjs';
-
-const isProd = (process.env.BUILD === 'production');
-
-const banner = 
-`/*
-THIS IS A GENERATED/BUNDLED FILE BY ROLLUP
-if you want to view the source visit the plugins github repository
-*/
-`;
-
-export default {
-  input: 'main.ts',
-  output: {
-    dir: '.',
-    sourcemap: 'inline',
-    sourcemapExcludeSources: isProd,
-    format: 'cjs',
-    exports: 'default',
-    banner,
-  },
-  external: [
-    "obsidian",
-    "electron",
-    "codemirror",
-    "@codemirror/autocomplete",
-    "@codemirror/closebrackets",
-    "@codemirror/commands",
-    "@codemirror/fold",
-    "@codemirror/gutter",
-    "@codemirror/history",
-    "@codemirror/language",
-    "@codemirror/rangeset",
-    "@codemirror/rectangular-selection",
-    "@codemirror/search",
-    "@codemirror/state",
-    "@codemirror/stream-parser",
-    "@codemirror/text",
-    "@codemirror/view",
-  ],
-  plugins: [
-    typescript(),
-    nodeResolve({browser: true}),
-    commonjs(),
-  ]
-};
